+++ R/java-shims.R
@@ -73,9 +73,7 @@ catch_exception <- function (object, ...) {
   # https://github.com/s-u/RJDBC/blob/1b7ccd4677ea49a93d909d476acf34330275b9ad/R/class.R#L18
   if (rJava::is.jnull(object)) {
     x <- rJava::.jgetEx(TRUE)
-    if (rJava::is.jnull(x))
-      stop(..., call. = FALSE)
-    else
+    if (!rJava::is.jnull(x))
       stop(..., ": ", rJava::.jcall(x, "S", "getMessage"), call. = FALSE)
   }
 }